+++ src/cli/roam.ts
@@ -10,6 +10,7 @@ import { createSaveCommand } from './commands/save.js';
 import { createRefsCommand } from './commands/refs.js';
 import { createUpdateCommand } from './commands/update.js';
 import { createBatchCommand } from './commands/batch.js';
+import { createRenameCommand } from './commands/rename.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -33,6 +34,7 @@ program.addCommand(createSaveCommand());
 program.addCommand(createRefsCommand());
 program.addCommand(createUpdateCommand());
 program.addCommand(createBatchCommand());
+program.addCommand(createRenameCommand());
 
 // Parse arguments
 program.parse();